Add task that makes LLVM dependency smaller
Currently Kotlin/Native uses default LLVM distribution.
While it is useful to have whole LLVM toolchain for development,
it is huge pain for users, because dependency is needlessly big.

The correct solution for this problem is to make own build script for LLVM,
but it is a big task that make take a while.

So for now we repack LLVM dependency without unnecessary files.